+++ CHANGELOG.md
@@ -21,6 +21,11 @@ Changes are grouped as follows
 
 ## Unreleased
 
+## [1.4.13] - 2020-03-12
+### Added
+- DataSets API and support for this in assets, events, time series, files and sequences.
+- .asset helper function on time series.
+
 ## [1.4.13] - 2020-03-03
 ### Added
 - Relationship list supports multiple sources, targets, relationship types and datasets.